requirements.txt
flask
pandas
requests
schedule
python-dotenv
supabase
sendgrid
gunicorn
scikit-learn
numpy
